+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project are documented in this file.
 - Added editing functionality for workspaces in `Workspaces.js` component.
 ### Changed
 - Refactored workspace name handling in `Workspaces.js` to improve variable naming and initialization for workspace creation and editing functions.
+- Updated `Workspaces.js` to switch from optimistic to pessimistic local merge approach.
 
 
 ## 2025-07-04